Rename device.py to queue_item.py
I also had to fix up the documentation of the QueueItem attributes
to use `#:` comments instead of an Attributes section of the class
docstring, because Sphinx was complaining that they were documented
twice.
