--- doc/katgpucbf.device.rst
@@ -1,7 +1,0 @@
-katgpucbf.device module
-=======================
-
-.. automodule:: katgpucbf.device
-   :members:
-   :undoc-members:
-   :show-inheritance:
