+++ doc/katgpucbf.queue_item.rst
@@ -1,0 +1,7 @@
+katgpucbf.queue_item module
+===========================
+
+.. automodule:: katgpucbf.queue_item
+   :members:
+   :undoc-members:
+   :show-inheritance:
