+++ doc/katgpucbf.rst
@@ -17,9 +17,9 @@ Submodules
 .. toctree::
    :maxdepth: 4
 
-   katgpucbf.device
    katgpucbf.meerkat
    katgpucbf.monitor
+   katgpucbf.queue_item
    katgpucbf.recv
    katgpucbf.ringbuffer
    katgpucbf.spead
