+++ src/katgpucbf/queue_item.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 ################################################################################
 
-"""Utilities for working with katsdpsigproc devices."""
+"""Provide :class:`QueueItem`."""
 
 from typing import List
 
@@ -40,17 +40,12 @@ class QueueItem:
     :meth:`enqueue_wait_for_events` or :meth:`async_wait_for_events`.
     This indicates that the operation is complete and the next thing can be
     done with whatever data is in that region of memory.
-
-    Attributes
-    ----------
-    timestamp
-        Timestamp of the item.
-    events
-        A list of GPU event markers generated by an
-        :class:`~katsdpsigproc.abc.AbstractCommandQueue`.
     """
 
+    #: Timestamp of the item.
     timestamp: int
+    #: A list of GPU event markers generated by an
+    #: :class:`~katsdpsigproc.abc.AbstractCommandQueue`.
     events: List[AbstractEvent]
 
     def __init__(self, timestamp: int = 0) -> None:
